Extract SpringModelUtils
